pyproject.toml
[project]
name = "dce"
version = "0.1.0"
description = "Dynamic Context Evolution for Scalable Synthetic Data Generation"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "chromadb>=0.4",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "hdbscan>=0.8",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyyaml>=6.0",
    "tenacity>=8.0",
    "python-dotenv>=1.0",
    "tqdm>=4.65",
    "jsonlines>=4.0",
]

[project.optional-dependencies]
downstream = [
    "transformers>=4.30",
    "datasets>=2.14",
    "torch>=2.0",
    "evaluate>=0.4",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
